Rename telemetry sample and update readme
