+++ python/samples/demos/telemetry/.env.example
@@ -1,4 +1,4 @@
 TELEMETRY_SAMPLE_CONNECTION_STRING="..."
-OTLP_ENDPOINT="..."
+OTLP_ENDPOINT="http://localhost:4317/"
 SEMANTICKERNEL_EXPERIMENTAL_GENAI_ENABLE_OTEL_DIAGNOSTICS=true
 SEMANTICKERNEL_EXPERIMENTAL_GENAI_ENABLE_OTEL_DIAGNOSTICS_SENSITIVE=true
