+++ python/samples/demos/telemetry/demo_plugins.py
