+++ python/samples/demos/telemetry/telemetry_sample_settings.py
